Removed unwanted logs

+++ nodeCRUD/controller/userController.js
@@ -1,17 +1,18 @@
 const asyncHandler = require("express-async-handler");
 const users = require("../models/userModels");
 const mongoose = require("mongoose");
+
 //@desc get all users
 //@route api/
-//@access public
+//@access private
 const getUsers = asyncHandler(async (req, res) => {
-  const getUsers = await users.find({});
+  const getUsers = await users.find({ user_id: req.user.id });
   res.status(200).json(getUsers);
 });
 
 //@desc create users
 //@route api/
-//@access public
+//@access private
 const createUser = asyncHandler(async (req, res) => {
   const { firstname, lastname, email, isActive } = req.body;
   if (!firstname || !lastname || !email) {
@@ -23,6 +24,7 @@ const createUser = asyncHandler(async (req, res) => {
     lastname,
     email,
     isActive,
+    user_id: req.user.id,
   });
   console.log(req.body);
   res.status(201).json(createusers);
@@ -30,25 +32,43 @@ const createUser = asyncHandler(async (req, res) => {
 
 //@desc update users
 //@route api/updateUser/id
-//@access public
+//@access private
 const updateUser = asyncHandler(async (req, res) => {
+  const getById = await users.findById(req.params.id);
+  console.log("req user id", req.user.id);
+  if (req.user.id !== getById.user_id.toString()) {
+    res.status(403);
+    throw new Error(
+      "You don't have permission to update another user user is "
+    );
+  }
   const updateuser = await users.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
+
   res.json(updateuser);
 });
 
 //@desc delete users
 //@route api/deleteUser
-//@access public
+//@access private
 const deleteUser = asyncHandler(async (req, res) => {
+  const getById = await users.findById(req.params.id);
+
+  if (getById.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "You don't have permission to update another user user is "
+    );
+  }
   const deleteuser = await users.findByIdAndDelete(req.params.id);
+
   res.status(200).send(deleteuser);
 });
 
 //@desc get user by id
 //@route api/
-//@access public
+//@access private
 const getUserByID = asyncHandler(async (req, res) => {
   const getById = await users.findById(req.params.id);
   res.status(200).json(getById);
@@ -56,7 +76,7 @@ const getUserByID = asyncHandler(async (req, res) => {
 
 //@desc get user by id
 //@route api/search/
-//@access public
+//@access private
 const searchUser = asyncHandler(async (req, res) => {
   const { email, id } = req.query;
   const matchQuery = {};
@@ -81,31 +101,6 @@ const searchUser = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("No user found with specified criteria!");
   }
-
-  // if (email) {
-  //   const getByEmail = await users.findOne({ email });
-  //   if (getByEmail) {
-  //     res.status(200).json(getByEmail);
-  //   } else {
-  //     res
-  //       .status(404)
-  //       .json({ message: "User with the provided email not found." });
-  //   }
-  // } else if (id) {
-  //   const getById = await users.findById(id);
-  //   if (getById) {
-  //     res.status(200).json(getById);
-  //   } else {
-  //     res.status(404).json({ message: "User with the provided ID not found." });
-  //   }
-  // } else if (isActive) {
-  //   const getIsActive = await users.find({ isActive });
-  //   if (getIsActive) {
-  //     res.status(200).json(getIsActive);
-  //   } else {
-  //     res.status(404).json({ message: "User with the provided ID not found." });
-  //   }
-  // }
 });
 
 module.exports = {
@@ -114,6 +109,5 @@ module.exports = {
   updateUser,
   deleteUser,
   getUserByID,
-  // getUserByEmail,
   searchUser,
 };
